Major moving of things.
This pull request attempts to further separate the stateful elements from
the functional elements of the code, and have those stateful elements be
concerned only with state they modify. One defining instance of this is to
move the encoder/decoder state out of the PACFile object: the class that
handles opening/reading/writing the compressed file is not repsonible for
performing the encoding as well -- instead a separate coding pipeline was
built to hand state off between the different steps in the encoder.

Additionally, this adds documentation to the README.md and removes some
unnecessary files.